pyproject.toml
[tool.poetry]
name = "aidungeon"
version = "0.1.0"
description = ""
authors = ["Nikolay Bryskin <nb@gosu.ai>"]

[tool.poetry.dependencies]
python = "3.7.*"
pyyaml = "^5.3.1"
profanityfilter = "^2.0.6"
tensorflow-gpu = "1.15.3"
aiohttp = "^3.6.2"
regex = "^2020.7.14"
tracery = "^0.1.1"

[tool.poetry.dev-dependencies]
gosu-tasks = {git = "https://github.com/gosuai/gosu-tasks"}
invoke = "^1.4.1"
flake8 = "^3.8.3"
ipython = "^7.18.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
